pyproject.toml
[project]
name = "csv-to-parquet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "polars>=0.20.0", 
    "pyarrow>=14.0.0",
]

[project.optional-dependencies]
dev = [
    "ipython>=8.18.0",
    "jupyter>=1.0.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
sources = ["src"]
